+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Expiry date has to be in the future
+
 ### 1.2.1: July 15th, 2019
 * Inform customers to come 1 hour early to enjoy Aquatonic
 * Avoid PHP notices and warnings